Perfect! Except that we’re still telling the client that the content is written in HTML. That’s the job of the Content-Type response header, and it defaults to text/html. Our response is being dishonest right now, and we risk confusing an API client.
Fix this by manually setting the Content-Type header on the Response before returning it.

How you set a response header may be different in your app, but there is definitely a way to do this. And because the Content-Type header is so important, you may even have a shortcut method for it.

Note
For example, in Laravel, you can return a JSON response (with correct Content-Type) with:
Response::json(array('name' => 'Steve', 'state' => 'CA'));
We’re now returning a JSON representation of the resource, setting its Content-Type header correctly and using the right status code. Great work.