Hey, we have a working endpoint to view a single programmer! We’re awesome :) Now do you remember the Location response header we return after creating a new programmer? Let’s update that to be a real value.
To do this, first add a bind function to our programmer route.

This gives the route an internal name of api_programmers_show. We can use that below to generate a proper URL to the new programmer resource.

The generateUrl method is a shortcut I added for our app, and it combines the nickname with the rest of the URL. You may make URLs differently in your app, but the idea is the same: set the Location header to the URI where I can GET this new resource.

Tip
The generateUrl method is just a shortcut for doing this:
{dollar_sign}programmerUrl = {dollar_sign}this->container['url_generator']->generate(
  'api_programmers_show',
  ['nickname' => {dollar_sign}programmer->nickname]
);
Update the testing.php script to print out the response from the original POST so we can check this out.
And when we run it again, we’ve got a working Location header:

HTTP/1.1 201 Created
...
Location: /api/programmers/ObjectOrienter330

It worked. Believe me - I'm an API