Using the Location Header
The Location header is more than just a nice thing. Its purpose is to help the client know where to go next without needing to hardcode URLs or URL patterns. To prove this, we can update our testing script to read the Location header and use it for the next request. This lets us remove the hardcoded URL pattern we had before.
If the URL pattern to view a programmer changes in the future, our client won’t break. That’s really powerful. But it’s also where things start to get complicated. More on that later, dear warrior.