We now have 2 URLs and so 2 resources:
  /api/programmers, which represents a collection of resources (i.e. all programmers);
  /api/programmers/{nickname}, which represents one programmer.
Actually, we can POST to the /api/programmers resource, but we can’t GET it yet. And nothing says that we have to support the GET method for a resource. But we’ll add it for two reasons. First, I’ll pretend that our imaginary iPhone app needs it. And second, API users tend to assume that you can GET most any resource. If we make this possible, our API is that much more predictable and friendly.