Fixing the Content-Type on POST Check us out! We now have 3 working endpoints, but one has a big issue. The POST still returns a text string as its response. Even if you don’t know what it should return, that’s embarassing. Come on, we can do better!
After creating a resource, one great option is to return a representation of the new resource. Use the serializeProgrammer to get JSON and put it into the Response.

If the client needs that information, then we’ve just saved them one API request.

And don’t forget to set the Content-Type response header to application/json. To see your handy work, print out that response temporarily and try it.