You guys know the drill: we start by writing the test. So let’s add yet another scenario.
This looks a lot like our POST scenario, and that’s good: consistency! But we do need to add a line to put a programmer into the database and tweak a few other details. The status code is different: 201 is used when an asset is created but the normal 200 is used when it’s an update.

Just to keep us tied into the theory of things, I’ll describe this using REST-nerd language. Ready? Ok.

This tests that when we send a 'representation' of a programmer resource via PUT, the server will use it to update that resource and return a representation.

We haven’t actually coded this yet, so when we run the test, it fails:

php vendir/bin/behat
The test reports that the status code isn’t 200, it’s 405. 405 means 'method not allowed', and our framework is doing this for us. It’s a way of saying 'Hey, /api/programmers/CowboyCoder is a valid URI, but the resource doesn’t support the PUT method.'

If your API doesn’t support an HTTP method for a resource, you should return a 405 response. If you use a decent framework, you should get this functionality for free.