Our tests are passing, but we’re doing a bad job: I’d created duplicated code in newAction and updateAction!
Let’s redeem ourselves! Create a new private function called handleRequest and copy the code into it that reads the request body and sets the data on the Programmer. Cool! Now we can just call this from newAction and updateAction.

Re-run the tests to see if we broke anything:

php vendir/bin/behat