GET Representation != POST Representation So far, the representation of a programmer that we send in our PUT request exactly matches the representation the server sends in the response after a GET request. But it does’t need to be this way.
When we POST or PUT to a programmer URI, we send data that contains an avatarNumber field. But imagine if we designed the API so that when we GET a programmer, the representation in the response contains an avatarUrl instead of avatarNumber. In this world, the representation we send in a request would be different than the representation that the server sends back in a response. And this would be perfectly ok!

The point is that I don’t want you to feel like the data your API receives needs to look exactly like the data you send back. Nope. Both are just representations of a resource.

With that said, if you make the representations inconsistent for no reason, your API users will hunt you down with pitchforks. So if you expect avatarNumber in the POST body, don’t send back avatar_number in the GET request. That’s just mean.