Handling Errors Things are looking nice, but actually, our API is kinda unusable.
We’re missing a really important piece: errors! Unless we properly handle errors, our API is going to be a pain to use because the client won’t know what it’s doing wrong. If we deployed right now and a client tried to create a programmer with a nickname that already exists, it would get a 500 error with no details. Bummer!