Let’s start with a simple case: when a client POST’s to /api/programmers, we should add validation to make the nickname field required.
Like always, let’s start with the test.

We can copy the working scenario for creating a programmer, but remove the nickname field from the request payload. Obviously, the status code won’t be 201. Because this is a client error, we’ll need a 400-level status code. But which one? Ah, this is another spot of wonderful debate! The most common is probably 400, which means simply 'Bad Request'. If we look back at the RFC 2616(http://www.w3.org/Protocols/rfc2616/rfc2616.html) document, the description of the 400 status code seems to fit our situation:

  The request could not be understood by the server due to malformed syntax. The client SHOULD NOT repeat the request without modifications.

The other common choice is 422: Unprocessable Entity. 422 comes from a different RFC and is used when the format of the data is ok, but semantically, it has errors. We’re adding validation for the business rule that a nickname is required, which is totally a semantic detail. So even though 422 seems to be less common than 400 for validation errors, it may be a more proper choice.

Now, what should the response content look like? Obviously, it’ll be JSON for our API, but let me suggest a structure that looks like this:

{
  'type': 'validation_error',
  'title': 'There was a validation error',
  'errors': {
     'nickname': 'Please enter a nickname'
   }
}
Just trust me on the structure for now. In the test, let’s look for these 3 fields and make sure we have a nickname error but no avatarNumber error