Test, check! Next, let’s actually add some validation. This is one of those spots that will be different based on your framework, but the important thing is how we communicate those errors to the client.
In Silex, to make the nickname field required, we need to open up the Programmer class itself. Here, add a NotBlank annotation with a nice message.
Cool! Next, open up the ProgrammerController class. In newAction, we should check the validation before saving the new Programmer. I’ve created a shortcut method called validate that does this.

It uses the annotation we just added to the Programmer class and returns an array of errors: one error for each field. If you are using Silex or Symfony, you can re-use my shortcut code with your project. If you’re not, just make sure you have some way of getting back an array of errors.

If the  array isn’t empty, we’ve got a problem! And since we already wrote the test, we know how we want to tell the user. Create an array with the type, title and errors fields. The  variable is already an associative array of messages, where the keys are the field names. So we can just set it on the errors field.
Just like with any other API response, we can create a JsonResponse class and pass it our data. The only difference with this endpoint is that it has a status code of 400.