What’s that? You want to check our validation rules when updating too? Great idea!
To avoid duplication, create a new private function in the controller called handleValidationResponse. We’ll pass it an array of errors and it will transform into living robotic beings originating from the distant machine world of Cybertron err ... I mean the proper 400 JSON response.
Now that we have this, use it in newAction and updateAction.
To make sure we didn’t break anything, we can run our tests:

We could have also added another scenario to test the validation when updating. But I’m not made of scenarios people! How detailed you get with your tests is up to you.